requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv
couchbase